requirements.txt
camelot-py==1.0.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
et_xmlfile==2.0.0
ghostscript==0.7
heroes==1.0.2
numpy==2.2.2
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
pandas==2.2.3
pdfminer.six==20240706
pycparser==2.22
pypdf==5.2.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pytz==2025.1
setuptools==75.8.0
six==1.17.0
tabulate==0.9.0
tzdata==2025.1
